raspi/requirements.txt
requests
pymongo[srv]
python-dotenv
RPi.GPIO
w1thermsensor